requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pillow==10.4.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
torch==2.4.1
torchvision==0.19.1
transformers==4.44.2
accelerate==0.34.2
ultralytics==8.3.0
pyyaml==6.0.2
opencv-python-headless==4.10.0.84
kagglehub==0.3.3
